requirements.txt
streamlit
xrpl-py
